Use static server id
Use a static server name defined in the settings file so plugins can be
configured to refer to a specific irc network in the settings. In
addition to this two new functions are added for plugins, `onconnect`
and `ondisconnect`. These functions notify the plugin that a specific
irc network was connected to or disconnected from. The name is defined
in the `name` attribute in the server config.

The autojoin function was moved to a separate plugin. Configure the
plugin like this:

{
    "freenode": [
        "#weechat",
        "#chromium"
    ]
}