Predicter: Adding a predicter plugin (aka Kuvert)
This can be used to make predictions about the future.  So, for example,
you can make a prediction that candidate X will win the US presidency in
2028 by sending a message to the bot like this:

!predict #channel 2028-11-08 02:10 Candidate X will win!

The prediction will be kept secret until the specified time and date.

The time of the prediction and who made the prediction will be made
public once the time and date has been reached.